refactor(agent): extract DAG graph building from execution supervisor
